backend/requirements.txt
asgiref==3.5.2
pycoingecko==3.1.0
Django==4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
setuptools==65.0.0
sqlparse==0.4.2
psycopg2-binary
PyJWT==2.6.0
wheel==0.36.2